pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pan-ai-security"
version = "0.1.0"
description = "Unified Python SDK for Palo Alto Networks AI Security — AIRS Runtime + WildFire"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Scott Thornton", email = "scthornton@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["palo-alto", "ai-security", "wildfire", "airs", "malware", "prompt-injection"]
dependencies = [
    "pan-aisecurity>=0.4.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "aioresponses>=0.7",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/scthornton/panw-unified-sdk"
Repository = "https://github.com/scthornton/panw-unified-sdk"
Issues = "https://github.com/scthornton/panw-unified-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pan_ai_security"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
